attemppt to fix weird comparison

+++ app/controllers/src/Orbit/Controller/API/v1/Pub/PromoCode/Repositories/PromoCodeRepository.php
@@ -73,7 +73,7 @@ class PromoCodeRepository implements RepositoryInterface
             'promo_code' => OrbitInput::post('promo_code'),
             'object_id' => OrbitInput::post('object_id'),
             'object_type' => OrbitInput::post('object_type'),
-            'quantity' => OrbitInput::post('qty'),
+            'quantity' => (int) OrbitInput::post('qty'),
 
             //is_final_check = true for final test before creating payment transaction
             'is_final_check' => OrbitInput::post('is_final_check'),
